+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.5] - 2026-03-12
+### Added
+- Introduced new keywords (`last`, `prev`, `previous`, `above`, and `_`) that reference the result of the preceding non-empty line.
+
 ## [2.4.4] - 2026-03-11
 ### Added
 - Added support to swipe left on any file to quickly delete it, with an undo option.